# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Logiflow API
  description: API для логистической информационной системы
  version: 0.2.0
  contact:
    name: anxi0uz
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
...
    DriverCreate:
      type: object
      required: [email, password, fullName, licenseNumber, licenseExpiry]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        vehicleId:
          type: string
          format: uuid
          nullable: true
        licenseNumber:
          type: string
        licenseExpiry:
          type: string
          format: date
    DriverUpdate:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        licenseExpiry:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [available, on_route, off_duty]
          nullable: true

    # ─── Order ────────────────────────────────────────────────────────────────
...
    AssignmentCreate:
      type: object
      required: [driverId, vehicleId, plannedFrom, plannedTo]
      properties:
        driverId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        plannedFrom:
          type: string
          format: date-time
        plannedTo:
          type: string
          format: date-time
        replacesAssignmentId:
          type: string
          format: uuid
          nullable: true
        source:
          type: string
          enum: [manual, dispatch_recommendation]
          default: manual
        recommendationId:
          type: string
          nullable: true
        recommendationCreatedAt:
          type: string
          format: date-time
          nullable: true
    AssignmentReject:
      type: object
      required: [reasonCode]
      properties:
        reasonCode:
          type: string
        comment:
          type: string
          nullable: true
    DeliveryComplete:
      type: object
      properties:
        recipientName:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
...
    DriverStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available, on_route, off_duty]

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      summary: Список водителей
      tags: [Drivers]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [available, on_route, off_duty]
      responses:
        "200":
          description: Список водителей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    post:
      operationId: createDriver
      summary: Создать водителя
      description: Только для роли admin. Создаёт пользователя с ролью driver и профиль водителя в одной транзакции.
      tags: [Drivers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverCreate"
      responses:
        "200":
          description: Водитель создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
  /drivers/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getDriver
      summary: Получить водителя по slug
      tags: [Drivers]
      responses:
        "200":
          description: Данные водителя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Не найдено

    put:
      operationId: updateDriver
      summary: Обновить данные водителя
      tags: [Drivers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Данные обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    delete:
      operationId: deleteDriver
      summary: Удалить водителя
      tags: [Drivers]
      responses:
        "200":
          description: Водитель удалён
  /drivers/me/status:
    patch:
      operationId: updateMyDriverStatus
      summary: Водитель меняет свой статус
      tags: [Drivers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverStatusUpdate"
      responses:
        "200":
          description: Статус обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Доступно только водителям

  # ─── Orders ─────────────────────────────────────────────────────────────────
...
  /api/v1/orders/{id}/assignments:
    post:
      operationId: createAssignment
      summary: Предложить или заменить назначение
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentCreate" }
      responses:
        "201": { description: Назначение создано }
        "409": { description: Ресурс занят или состояние изменилось }
  /api/v1/assignments/{id}/accept:
    post:
      operationId: acceptAssignment
      summary: Принять назначение
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Назначение принято }
  /api/v1/assignments/{id}/reject:
    post:
      operationId: rejectAssignment
      summary: Отклонить назначение
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentReject" }
      responses:
        "200": { description: Назначение отклонено }
  /api/v1/assignments/{id}/start:
    post:
      operationId: startAssignment
      summary: Начать перевозку
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Перевозка началась }
  /api/v1/assignments/{id}/arrive:
    post:
      operationId: arriveAssignment
      summary: Подтвердить прибытие
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Прибытие подтверждено }
  /api/v1/assignments/{id}/complete:
    post:
      operationId: completeAssignment
      summary: Подтвердить доставку
      tags: [Assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeliveryComplete" }
      responses:
        "200": { description: Доставка завершена }
...